refactor: moved middleware to own namespace
