+++ src/Xenial.Tasty/Execution/TestDelegate.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Xenial.Delicious.Execution
+{
+    public delegate Task TestDelegate(TestContext context);
+}
