+++ src/Xenial.Tasty/Execution/TestExecutorExtentions.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Xenial.Delicious.Execution
+{
+    public static class TestExecutorExtentions
+    {
+        public static TestExecutor Use(this TestExecutor app, Func<TestContext, Func<Task>, Task> middleware)
+        {
+            return app.Use(next =>
+            {
+                return context =>
+                {
+                    Func<Task> simpleNext = () => next(context);
+                    return middleware(context, simpleNext);
+                };
+            });
+        }
+    }
+}
